instagram-demo/requirements.txt
Appium-Python-Client
selenium
pytest

# pip install -r requirements.txt